+++ tests/test_conway_continuity.h
@@ -804,7 +804,7 @@ inline void test_collapsing_faces_land_on_host_palette() {
 
     // Bookend: the hankin star-face classification of the arrival node, as
     // start_morph_cycle builds it.
-    int arrival_topo[MAX_FACES];
+    uint16_t arrival_topo[MAX_FACES];
     {
       Arena ha(cc_scan_buf, sizeof(cc_scan_buf) / 2);
       Arena hb(cc_scan_buf + sizeof(cc_scan_buf) / 2, sizeof(cc_scan_buf) / 2);
@@ -832,8 +832,8 @@ inline void test_collapsing_faces_land_on_host_palette() {
     for (size_t f = survivors; f < landing.faces; ++f) {
       bool hosted = false;
       for (size_t j = 0; j < survivors && !hosted; ++j)
-        hosted = wrap(landing.topology[f], PALETTES) ==
-                 wrap(landing.topology[j], PALETTES);
+        hosted = wrap(static_cast<int>(landing.topology[f]), PALETTES) ==
+                 wrap(static_cast<int>(landing.topology[j]), PALETTES);
       HS_EXPECT_TRUE(hosted);
     }
   }
@@ -901,9 +901,8 @@ inline void test_crossfade_exact_at_endpoints_emission() {
   step_and_snapshot(anim, fx, snap);
   HS_EXPECT_EQ(snap.colors.size(), landing.faces);
   for (size_t f = 0; f < snap.colors.size(); ++f) {
-    const uint8_t to =
-        landing
-            .to_palette[wrap(landing.topology[f], Animation::OpLeg::PALETTES)];
+    const uint8_t to = landing.to_palette[wrap(
+        static_cast<int>(landing.topology[f]), Animation::OpLeg::PALETTES)];
     const uint8_t from = f < landing.primary_faces
                              ? pal[f]
                              : to; // newborn faces skip the crossfade
@@ -916,9 +915,8 @@ inline void test_crossfade_exact_at_endpoints_emission() {
   for (int f = 1; f < SWEEP; ++f)
     step_and_snapshot(anim, fx, snap);
   for (size_t f = 0; f < snap.colors.size(); ++f) {
-    const uint8_t to =
-        landing
-            .to_palette[wrap(landing.topology[f], Animation::OpLeg::PALETTES)];
+    const uint8_t to = landing.to_palette[wrap(
+        static_cast<int>(landing.topology[f]), Animation::OpLeg::PALETTES)];
     for (int s = 0; s < NUM_RAMP_SAMPLES; ++s)
       expect_color_eq(snap.colors[f][s],
                       bank.bank.entries[to].get(RAMP_SAMPLES[s]));
@@ -1068,8 +1066,8 @@ inline void test_palette_mapping_total_all_edges() {
     step_and_snapshot(anim, fx, snap); // frame 1: w == 0
     HS_EXPECT_EQ(snap.colors.size(), landing.faces);
     for (size_t f = 0; f < snap.colors.size(); ++f) {
-      const uint8_t to = landing.to_palette[wrap(landing.topology[f],
-                                                 Animation::OpLeg::PALETTES)];
+      const uint8_t to = landing.to_palette[wrap(
+          static_cast<int>(landing.topology[f]), Animation::OpLeg::PALETTES)];
       const uint8_t from = f < landing.primary_faces ? pal[f] : to;
       for (int s = 0; s < NUM_RAMP_SAMPLES; ++s)
         expect_color_eq(snap.colors[f][s],
